+++ middleware/wifi_nxp/incl/wlcmgr/wlan_11d.h
@@ -1,23 +1,7 @@
 /*
- *  Copyright 2008-2020 NXP
+ *  Copyright 2008-2022 NXP
  *
- *  NXP CONFIDENTIAL
- *  The source code contained or described herein and all documents related to
- *  the source code ("Materials") are owned by NXP, its
- *  suppliers and/or its licensors. Title to the Materials remains with NXP,
- *  its suppliers and/or its licensors. The Materials contain
- *  trade secrets and proprietary and confidential information of NXP, its
- *  suppliers and/or its licensors. The Materials are protected by worldwide copyright
- *  and trade secret laws and treaty provisions. No part of the Materials may be
- *  used, copied, reproduced, modified, published, uploaded, posted,
- *  transmitted, distributed, or disclosed in any way without NXP's prior
- *  express written permission.
- *
- *  No license under any patent, copyright, trade secret or other intellectual
- *  property right is granted to or conferred upon you by disclosure or delivery
- *  of the Materials, either expressly, by implication, inducement, estoppel or
- *  otherwise. Any license under such intellectual property rights must be
- *  express and approved by NXP in writing.
+ *  Licensed under the LA_OPT_NXP_Software_License.txt (the "Agreement")
  *
  */
 
@@ -116,7 +100,7 @@ static inline int wlan_get_country(void)
  */
 static inline int wlan_uap_set_country(country_code_t country)
 {
-    return wifi_uap_set_country(country);
+    return wifi_uap_set_country((int)country);
 }
 
 /** Set country code in WLAN Driver.
@@ -136,7 +120,7 @@ static inline int wlan_uap_set_country(country_code_t country)
  */
 static inline int wlan_set_country(country_code_t country)
 {
-    return wifi_set_country(country);
+    return wifi_set_country((int)country);
 }
 
 /**  wlan_11d_custom Custom Wi-Fi Region Configuration
