+++ middleware/wifi_nxp/wifidriver/mlan_sta_cmdresp.c
@@ -4,23 +4,7 @@
  *
  *  Copyright 2008-2022 NXP
  *
- *  NXP CONFIDENTIAL
- *  The source code contained or described herein and all documents related to
- *  the source code ("Materials") are owned by NXP, its
- *  suppliers and/or its licensors. Title to the Materials remains with NXP,
- *  its suppliers and/or its licensors. The Materials contain
- *  trade secrets and proprietary and confidential information of NXP, its
- *  suppliers and/or its licensors. The Materials are protected by worldwide copyright
- *  and trade secret laws and treaty provisions. No part of the Materials may be
- *  used, copied, reproduced, modified, published, uploaded, posted,
- *  transmitted, distributed, or disclosed in any way without NXP's prior
- *  express written permission.
- *
- *  No license under any patent, copyright, trade secret or other intellectual
- *  property right is granted to or conferred upon you by disclosure or delivery
- *  of the Materials, either expressly, by implication, inducement, estoppel or
- *  otherwise. Any license under such intellectual property rights must be
- *  express and approved by NXP in writing.
+ *  Licensed under the LA_OPT_NXP_Software_License.txt (the "Agreement")
  *
  */
 
@@ -228,7 +212,7 @@ static mlan_status wlan_get_power_level(pmlan_private pmpriv, void *pdata_buf)
         {
             max_power = pg->power_max;
             min_power = pg->power_min;
-            length -= sizeof(Power_Group_t);
+            length -= (int)sizeof(Power_Group_t);
         }
         while (length > 0)
         {
@@ -241,7 +225,7 @@ static mlan_status wlan_get_power_level(pmlan_private pmpriv, void *pdata_buf)
             {
                 min_power = pg->power_min;
             }
-            length -= sizeof(Power_Group_t);
+            length -= (int)sizeof(Power_Group_t);
         }
         if (ppg_tlv->length > 0U)
         {
@@ -416,6 +400,7 @@ static mlan_status wlan_ret_802_11_rf_channel(IN pmlan_private pmpriv,
     LEAVE();
     return MLAN_STATUS_SUCCESS;
 }
+
 
 
 /********************************************************
